Final project VO
This WIP code uses SIFT or ORB descriptors to perform VO. it works on simple trajectories and in limited cases where keypoints are easy enough to aquire but not on more complex or challenging scenes. I think this is where we can pull in some deep learning and get more robust performance.